+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
-* `Contact` now has `is_emailer`, `is_interesting`, and `link_to_bio` props and `bio()` method for biographical panel
+* `DocCfg` add `date_uncertain`, `is_valid_for_name_scan`, and `non_participants` fields
+* `Contact` now has `is_emailer`, `is_interesting`, `match_partial_names`, and `link_to_bio` props and `bio()` method for biographical panel
 
 ### 1.8.7
 * More biographical information panels in chrono view